add: nanoid package, enum for loading status, interfaces.ts;
