+++ package.json
@@ -13,6 +13,7 @@
     "@types/react-dom": "^17.0.13",
     "axios": "^0.26.1",
     "classnames": "^2.3.1",
+    "nanoid": "^3.3.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-hook-form": "^7.28.0",
