+++ src/components/Sidebar/Sidebar.tsx
@@ -3,6 +3,7 @@ import spinner from '../../image/spinner.gif';
 import { Tag } from './Tag';
 import { Link } from './Link';
 import { useFetch } from '../hooks/useFetch';
+import { TypeLoadingStatus } from '../../interfaces/interfaces';
 
 const links = [
   { id: 1, name: 'Help' },
@@ -17,16 +18,21 @@ const links = [
 ];
 
 export function Sidebar() {
-  const { data, loading } = useFetch<string[]>('/tags');
+  const { data, loading, error } = useFetch<string[]>('/tags');
   return (
     <div className="xl:right-[10%] lg:right-[5%] xl:w-1/4 lg:w-1/4 lg:fixed md:w-full md:static sm:w-full sm:static w-full flex flex-col">
       <p className="text-xs font-bold">DISCOVER MORE OF WHAT MATTERS TO YOU</p>
       <div className="h-24 w-full pt-4 pl-2 flex flex-wrap justify-start">
-        {loading ? (
+        {error && loading === TypeLoadingStatus.IS_REJECTED ? (
+          <div className="w-full flex justify-center items-center">
+            <p className="text-sm">Sorry, tag field is not available now!</p>
+          </div>
+        ) : loading === TypeLoadingStatus.IS_PENDING ? (
           <div className="w-full h-full flex justify-center items-center">
             <img src={spinner} width={20} height={20} alt="spinner" />
           </div>
         ) : (
+          loading === TypeLoadingStatus.IS_RESOLVE &&
           data?.map((tag, index) => <Tag key={index} tag={tag} />)
         )}
         <hr className="w-full mt-2 mb-2" />
