+++ src/components/hooks/useFetch.ts
@@ -1,41 +1,23 @@
 import { useState, useEffect } from 'react';
 
 import { fetchDataService } from '../../services/getResponseDataService';
-
-export interface IArticles {
-  slug: string;
-  title: string;
-  description: string;
-  body: string;
-  tagList: string[];
-  createdAt: string;
-  updatedAt: string;
-  favoritesCount: number;
-  author: IAuthor;
-}
-
-interface IAuthor {
-  username: string;
-  bio: string;
-  image: string;
-  following: boolean;
-}
+import { TypeLoadingStatus } from '../../interfaces/interfaces';
 
 export const useFetch = <T>(url: string) => {
   const [data, setData] = useState<T | null>(null);
-  const [loading, setLoading] = useState(false);
+  const [loading, setLoading] = useState<TypeLoadingStatus>(TypeLoadingStatus.IS_RESOLVE);
   const [error, setError] = useState(false);
   const getResponseData = (url: string) => {
-    setLoading(true);
+    setLoading(TypeLoadingStatus.IS_PENDING);
     fetchDataService(url)
       .then((res) => {
-        setLoading(false);
+        setLoading(TypeLoadingStatus.IS_RESOLVE);
         const urlKey = url.split('').splice(1, url.length).join('');
         setData(res.data[urlKey]);
       })
       .catch((err) => {
         console.log(err);
-        setLoading(false);
+        setLoading(TypeLoadingStatus.IS_REJECTED);
         setError(true);
       });
   };
